refactor: Refactor settings page
Massive refactor of settings page for configuring number generator sequences. SHOULD be feature-equivalent to what we had before, ready for the meat of the work on SI-4 to start

SI-4
