+++ src/public/hooks/index.js
@@ -1,4 +1,8 @@
 export { default as useNumberGenerators } from './useNumberGenerators';
+export { default as useNumberGeneratorSequences } from './useNumberGeneratorSequences';
+export { default as useNumberGeneratorSequence } from './useNumberGeneratorSequence';
+
 export { default as useMutateNumberGeneratorSequence } from './useMutateNumberGeneratorSequence';
 export { default as useMutateNumberGenerator } from './useMutateNumberGenerator';
+
 export { default as useGenerateNumber } from './useGenerateNumber';
