+++ src/public/utilities/index.js
@@ -1,7 +1,4 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as ModalButton } from './ModalButton';
 
-export {
-  NUMBER_GENERATORS_ENDPOINT,
-  NUMBER_GENERATOR_ENDPOINT
-} from './endpoints';
+export * from './endpoints';
