+++ src/settings/NumberGeneratorSequenceRoute/SequenceView.js
@@ -1,14 +1,19 @@
-import { useCallback } from 'react';
-
+import { useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+
+import { FormModal } from '@k-int/stripes-kint-components';
+
+import { useCallout } from '@folio/stripes/core';
 
 import {
   Button,
   Col,
+  ConfirmationModal,
   Icon,
   KeyValue,
   Pane,
+  PaneHeader,
   Row
 } from '@folio/stripes/components';
 
@@ -20,21 +25,59 @@ import {
   NameInfo,
   NextValueInfo,
   OutputTemplateInfo
-} from './InfoPopovers';
+} from '../InfoPopovers';
+
+import { useMutateNumberGeneratorSequence, useNumberGeneratorSequence } from '../../public';
+import NumberGeneratorSequenceForm from './NumberGeneratorSequenceForm';
 
 const NumberGeneratorSequence = ({
+  match: { params: { seqId } },
   onClose,
-  onDelete,
-  sequence,
-  setEditing,
 }) => {
+  const callout = useCallout();
+  const { data: sequence = {} } = useNumberGeneratorSequence({
+    id: seqId,
+    queryOptions: {
+      enabled: !!seqId
+    }
+  });
+
+  const [editing, setEditing] = useState(false);
+  const [showDeleteConfirmationModal, setShowDeleteConfirmationModal] = useState(false);
+
+  const {
+    put: editSeq,
+    delete: removeSeq
+  } = useMutateNumberGeneratorSequence({
+    afterQueryCalls: {
+      delete: () => {
+        callout.sendCallout({
+          message: <FormattedMessage
+            id="ui-service-interaction.settings.numberGeneratorSequences.callout.delete"
+            values={{ name: sequence?.name }}
+          />
+        });
+      },
+      put: (putValues) => {
+        setEditing(false);
+        callout.sendCallout({
+          message: <FormattedMessage
+            id="ui-service-interaction.settings.numberGeneratorSequences.callout.edit"
+            values={{ name: putValues?.sequences?.find(s => s.id === sequence.id)?.name }}
+          />
+        });
+      },
+    },
+    id: sequence?.owner?.id
+  });
+
   const actionMenu = useCallback(() => (
     [
       <Button
         key="action-edit-sequence"
         buttonStyle="dropdownItem"
         marginBottom0
-        onClick={setEditing}
+        onClick={() => setEditing(true)}
       >
         <Icon icon="edit">
           <FormattedMessage id="ui-service-interaction.edit" />
@@ -44,23 +87,28 @@ const NumberGeneratorSequence = ({
         key="action-delete-sequence"
         buttonStyle="dropdownItem"
         marginBottom0
-        onClick={onDelete}
+        onClick={() => setShowDeleteConfirmationModal(true)}
       >
         <Icon icon="trash">
           <FormattedMessage id="ui-service-interaction.delete" />
         </Icon>
       </Button>
     ]
-  ), [onDelete, setEditing]);
+  ), [setEditing]);
 
   return (
     <>
       <Pane
-        actionMenu={actionMenu}
         defaultWidth="fill"
-        dismissible
-        onClose={onClose}
-        paneTitle={sequence.name ?? sequence.code}
+        renderHeader={(renderProps) => (
+          <PaneHeader
+            {...renderProps}
+            actionMenu={actionMenu}
+            dismissible
+            onClose={onClose}
+            paneTitle={sequence.name ?? sequence.code}
+          />
+        )}
       >
         <Row>
           <Col xs={6}>
@@ -156,28 +204,59 @@ const NumberGeneratorSequence = ({
           </Col>
         </Row>
       </Pane>
+      {editing &&
+        <FormModal
+          initialValues={sequence}
+          modalProps={{
+            dismissible: true,
+            label: <FormattedMessage id="ui-service-interaction.settings.numberGeneratorSequences.editModal" />,
+            onClose: () => setEditing(false),
+            open: editing
+          }}
+          onSubmit={editSeq}
+        >
+          <NumberGeneratorSequenceForm />
+        </FormModal>
+      }
+      <ConfirmationModal
+        buttonStyle="danger"
+        confirmLabel={
+          <FormattedMessage id="ui-service-interaction.delete" />
+        }
+        heading={
+          <FormattedMessage id="ui-service-interaction.settings.numberGeneratorSequences.deleteSequence" />
+        }
+        message={[
+          <FormattedMessage
+            key="delete-sequence-message"
+            id="ui-service-interaction.settings.numberGeneratorSequences.deleteSequence.message"
+            values={{ name: sequence?.name }}
+          />,
+          <br key="delete-sequence-line-break" />,
+          <FormattedMessage
+            key="delete-sequence-warning-message"
+            id="ui-service-interaction.settings.numberGeneratorSequences.deleteSequence.warningMessage"
+          />
+        ]}
+        onCancel={() => setShowDeleteConfirmationModal(false)}
+        onConfirm={() => {
+          removeSeq(sequence?.id);
+          onClose();
+          setShowDeleteConfirmationModal(false);
+        }}
+        open={showDeleteConfirmationModal}
+      />
     </>
   );
 };
 
 NumberGeneratorSequence.propTypes = {
-  onClose: PropTypes.func.isRequired,
-  onDelete: PropTypes.func.isRequired,
-  sequence: PropTypes.shape({
-    code: PropTypes.string.isRequired,
-    checkDigitAlgo: PropTypes.shape({
-      label: PropTypes.string,
-      value: PropTypes.string
-    }),
-    description: PropTypes.string,
-    enabled: PropTypes.bool,
-    format: PropTypes.string,
-    id: PropTypes.string.isRequired,
-    name: PropTypes.string,
-    nextValue: PropTypes.number,
-    outputTemplate: PropTypes.string,
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      seqId: PropTypes.string
+    }).isRequired,
   }).isRequired,
-  setEditing: PropTypes.func.isRequired
+  onClose: PropTypes.func.isRequired,
 };
 
 export default NumberGeneratorSequence;
